+++ py-interface/__init__.py
@@ -109,6 +109,7 @@ from .amdsmi_interface import amdsmi_get_clock_info
 
 from .amdsmi_interface import amdsmi_get_pcie_info
 from .amdsmi_interface import amdsmi_get_gpu_bad_page_info
+from .amdsmi_interface import amdsmi_get_gpu_bad_page_threshold
 from .amdsmi_interface import amdsmi_get_violation_status
 from .amdsmi_interface import amdsmi_get_gpu_xgmi_link_status